4.21*: Blog List Expansion, step 9
Change the delete blog operation so that a blog can be deleted only by the user who added it. Therefore, deleting a blog is possible only if the token sent with the request is the same as that of the blog's creator.

If deleting a blog is attempted without a token or by an invalid user, the operation should return a suitable status code.

Note that if you fetch a blog from the database,
the field blog.user does not contain a string, but an object. So if you want to compare the ID of the object fetched from the database and a string ID, a normal comparison operation does not work. The ID fetched from the database must be parsed into a string first.

+++ part4/bloglist/controller/blogRouter.js
@@ -23,7 +23,7 @@ blogRouter.post("/", async (request, response, next) => {
     if (!decodedToken.id) {
       return response.status(401).json({ error: "token invalid" });
     }
-    
+
     const user = await User.findById(decodedToken.id);
 
     const blog = new Blog({
@@ -42,10 +42,21 @@ blogRouter.post("/", async (request, response, next) => {
   }
 });
 
-blogRouter.delete("/:id", async (request, response) => {
+blogRouter.delete("/:id", async (request, response, next) => {
   try {
-    await Blog.findByIdAndDelete(request.params.id);
-  } catch (error) {}
+    const decodedToken = jwt.verify(request.token, process.env.JWT_SECRET);
+
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: "token invalid" });
+    }
+
+    const blog = await Blog.findById(request.params.id);
+    if (blog.user.toString() === decodedToken.id.toString())
+      await Blog.findByIdAndDelete(request.params.id);
+    else response.status(403).json({ error: "token belongs to another user" });
+  } catch (error) {
+    next(error);
+  }
   response.status(204).end();
 });
